Crop off drop shadow
I think the drop shadows in the original screenshots were making for a
weird double-drop shadow in the combined gif. I banged my head against
the wall trying to figure out how to smooth it all out, but in the end I
just gave up and cropped off the bottom edge. Maybe this will look
better?