Fix TabLineSel highlighting
When I was generating screenshots for the wiki, I realized that the
default 'tabline' label generates a number in the tab name if you have
splits. This wouldn't be a problem, except that the number is
highlighted with the Title group, which I have as slate grey. So using
the slate background on TabLineSel renders the number invisible (the tab
label just looks like it gets inexplicably wider).

Instead of using the active selection highlighting for the selected tab,
I'll use the active UI highlighting. The contrast with Title isn't
great, but it's not totally invisible. And this also makes TabLineSel
and TabLine match StatusLine and StatusLineNC. Anyway, I won't mind
either way because I don't use tabs (and if I did, I'd probably set
guioptions+=e).